requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
aiofiles==23.2.1

# AI Models (CPU-only for Intel UHD)
torch>=2.0.0
torchvision>=0.15.0
gfpgan>=1.3.8
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0

# Image/Video Processing
opencv-contrib-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0

# Number Plate Detection & OCR
tensorflow>=2.10.0           # WPOD-NET deep learning framework
keras                        # WPOD-NET model parsing
ultralytics>=8.0.0           # YOLOv8 - plate bounding box detection
paddlepaddle                  # PaddleOCR CPU backend
paddleocr>=2.7.0             # OCR engine for text extraction
